+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `NormalizedName` for normalized environment variable names
+
 
 ## [0.2.0] - 2018-06-06